+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# weitrix 1.1.6
+
+* Use glm2, which is less prone to optimization failure.
+* Auto-disable parallel processing if X11 device is open.
+* Add SLAM-Seq vignette.
 
 # weitrix 1.1.5